+++ central-js/server/components/admin/index.js
@@ -1,0 +1,43 @@
+var NodeCache = require("node-cache" );
+var uuid = require("uuid");
+
+var adminKeysCache = new NodeCache();
+var cacheKey = 'admin-keys-map';
+// ИНН админов
+var aAdminInn = [
+	'3119325858'
+];
+
+var getAdminKeys = function () {
+	var result = adminKeysCache.get(cacheKey);
+	if (!result) {
+		result = {};
+		setAdminKeys(result);
+	}
+	return result;
+};
+
+var setAdminKeys = function (value) {
+	adminKeysCache.set(cacheKey, value);
+};
+
+var generateAdminToken = function (inn) {
+	var result = uuid.v1();
+	var keys = getAdminKeys();
+	keys[inn] = result;
+	setAdminKeys(keys);
+	return result;
+};
+
+var isAdminInn = function(inn) {
+	return aAdminInn.indexOf(inn) > -1;
+};
+
+var Admin = function() {
+
+};
+
+Admin.generateAdminToken = generateAdminToken;
+Admin.isAdminInn = isAdminInn;
+
+module.exports = Admin;
